in progress User and Item class reformat completed

+++ src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -16,27 +16,27 @@ public class ItemController {
     private final ItemService itemService;
 
     @GetMapping("/{itemId}")
-    public ItemDto getItem(@PathVariable Integer itemId) {
+    public ItemDto getItem(@PathVariable Long itemId) {
         log.debug(String.format("Item with id /%d", itemId));
         return itemService.getItem(itemId);
     }
 
     @GetMapping
-    public List<ItemDto> getItems(@RequestHeader("X-Sharer-User-Id") Integer userId) {
+    public List<ItemDto> getItems(@RequestHeader("X-Sharer-User-Id") Long userId) {
         log.debug(String.format("Items of user with id /%d ", userId));
         return itemService.getAllItems(userId);
     }
 
     @PostMapping
-    public ItemDto createItem(@RequestHeader("X-Sharer-User-Id") Integer userId,
+    public ItemDto createItem(@RequestHeader("X-Sharer-User-Id") Long userId,
                               @RequestBody @Valid ItemDto itemDto) {
         log.debug("Item created");
         return itemService.addItem(userId, itemDto);
     }
 
     @PatchMapping("/{itemId}")
-    public ItemDto updateItem(@RequestHeader("X-Sharer-User-Id") Integer userId,
-                              @PathVariable Integer itemId,
+    public ItemDto updateItem(@RequestHeader("X-Sharer-User-Id") Long userId,
+                              @PathVariable Long itemId,
                               @RequestBody ItemDto itemDto) {
         log.debug("Item updated");
         return itemService.updateItem(userId, itemId, itemDto);
@@ -49,7 +49,7 @@ public class ItemController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteItem(@PathVariable Integer id) {
+    public void deleteItem(@PathVariable Long id) {
         itemService.deleteItem(id);
         log.debug(String.format("Item with id /%d was deleted", id));
     }
